Move vectorize wrapper to vectorize_codegen
